Move IEndpointResponse implementation types to separate files
