+++ src/Atc.Rest.Client/EndpointResponse.cs
@@ -49,51 +49,4 @@ namespace Atc.Rest.Client
                    throw new InvalidCastException($"ContentObject is not of type {typeof(TResult).Name}");
         }
     }
-
-    public class EndpointResponse<T>
-        : EndpointResponse
-        where T : class
-    {
-        public EndpointResponse(EndpointResponse response)
-            : base(response)
-        {
-        }
-
-        public EndpointResponse(
-            bool isSuccess,
-            HttpStatusCode statusCode,
-            string content,
-            T? contentObject,
-            IReadOnlyDictionary<string, IEnumerable<string>> headers)
-            : base(isSuccess, statusCode, content, contentObject, headers)
-        {
-        }
-
-        public T? SuccessContent => IsSuccess ? CastContent<T>() : null;
-    }
-
-    public class EndpointResponse<TSuccessContent, TErrorContent>
-        : EndpointResponse
-        where TSuccessContent : class
-        where TErrorContent : class
-    {
-        public EndpointResponse(EndpointResponse response)
-            : base(response)
-        {
-        }
-
-        public EndpointResponse(
-            bool isSuccess,
-            HttpStatusCode statusCode,
-            string content,
-            object? contentObject,
-            IReadOnlyDictionary<string, IEnumerable<string>> headers)
-            : base(isSuccess, statusCode, content, contentObject, headers)
-        {
-        }
-
-        public TSuccessContent? SuccessContent => IsSuccess ? CastContent<TSuccessContent>() : null;
-
-        public TErrorContent? ErrorContent => !IsSuccess ? CastContent<TErrorContent>() : null;
-    }
 }
